# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cachet API",
    "description": "Issuance console and verifiable registry for Zcash Shielded Assets (ZSA), running against the public ZSA testnet.\n\nEverything the registry serves is **derived from the chain** and independently checkable: metadata hashes participate in the asset id itself (ZIP 227), so a response can be re-verified client-side. The public instance at `api.cachetzec.com` is **read-only** — every mutating `issuance` endpoint answers 403 there; browser mints go through `POST /api/v1/relay`, which only forwards signed bytes.\n\nRate limit: 30 requests/s per client, bursts of 60 - sized for a real session, a browser wallet paging through the chain included. Need better limits, or none at all? Run your own instance: the code is MIT, the limit is one environment variable (`CACHET_RATE_LIMIT_PER_SEC`, `0` disables it), and the public ZSA node accepts transactions directly. This deployment is a convenience, never a chokepoint.\n\nNo cookies, no telemetry, and no authentication anywhere except the optional operator surface.",
    "license": {
      "name": "MIT"
    },
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/chain": {
      "get": {
        "tags": ["chain"],
        "summary": "Identify the connected network and chain tip.",
        "operationId": "chain_info",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainInfoResponse"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChainInfoResponse": {
        "type": "object",
        "description": "Connected network, chain tip, and deployment mode.",
        "required": ["network", "tip_height", "read_only", "mints_paused"],
        "properties": {
          "mints_paused": {
            "type": "boolean",
            "description": "When true, the operator has paused minting through this instance:\nthe relay and metadata uploads answer 503 until it is lifted. The\nchain itself is unaffected."
          },
          "network": {
            "type": "string",
            "description": "Network name, e.g. `regtest`, `zsa-testnet`, `in-memory`.",
            "example": "regtest"
          },
          "read_only": {
            "type": "boolean",
            "description": "When true, the wallet-signing endpoints (mint, batch mint,\ntransfer, burn, wallet) answer 403 — a public read-only\ndeployment. Relay, description resolution and metadata upload stay\nopen: they need no key from this instance."
          },
          "snapshot_public_key": {
            "type": ["string", "null"],
            "description": "Ed25519 public key (hex) that signs this instance's registry\nsnapshots; absent when snapshots are not enabled. Compare it to\nthe operator's out-of-band publications (working paper, posts)\nbefore trusting a mirror."
          },
          "tip_height": {
            "type": "integer",
            "format": "int64",
            "description": "Chain tip height as seen by the node.",
            "minimum": 0
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 9457 problem document. Every non-2xx response uses this shape.",
        "required": ["type", "title", "status", "detail"],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code, duplicated for clients that drop it.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary."
          },
          "type": {
            "type": "string",
            "description": "Stable, documentation-friendly error identifier.",
            "example": "https://cachetzec.com/problems/asset-finalized"
          }
        }
      },
...
    }
  },
...
}